requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==3.7.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
openmeteo-requests
requests-cache
retry-requests
numpy